src/modelling_cogtran.py
from transformers import PretrainedConfig, PreTrainedModel
from transformers.modeling_outputs import TokenClassifierOutput
from esm.model.msa_transformer import MSATransformer
import torch
import torch.nn as nn
from typing import Optional, Union

class MSATConfig(PretrainedConfig):
    r"""
    This is the configuration class to store the configuration of a [`MSAModel`]. It is used to instantiate a MSA model
    according to the specified arguments, defining the model architecture. Instantiating a configuration with the
    defaults will yield a similar configuration to that of the MSA transformer.
    Configuration objects inherit from [`PretrainedConfig`] and can be used to control the model outputs. Read the
    documentation from [`PretrainedConfig`] for more information.
    Args:
        vocab_size (`int`, *optional*):
            Vocabulary size of the MSA model. Defines the number of different tokens that can be represented by the
            `inputs_ids` passed when calling [`MSAModel`].
        mask_token_id (`int`, *optional*):
            The index of the mask token in the vocabulary. This must be included in the config because of the
            "mask-dropout" scaling trick, which will scale the inputs depending on the number of masked tokens.
        pad_token_id (`int`, *optional*):
            The index of the padding token in the vocabulary. This must be included in the config because certain parts
            of the ESM code use this instead of the attention mask.
        hidden_size (`int`, *optional*, defaults to 768):
            Dimensionality of the encoder layers and the pooler layer.
        num_hidden_layers (`int`, *optional*, defaults to 12):
            Number of hidden layers in the Transformer encoder.
        num_attention_heads (`int`, *optional*, defaults to 12):
            Number of attention heads for each attention layer in the Transformer encoder.
        intermediate_size (`int`, *optional*, defaults to 3072):
            Dimensionality of the "intermediate" (often named feed-forward) layer in the Transformer encoder.
        hidden_dropout_prob (`float`, *optional*, defaults to 0.1):
            The dropout probability for all fully connected layers in the embeddings, encoder, and pooler.
        attention_probs_dropout_prob (`float`, *optional*, defaults to 0.1):
            The dropout ratio for the attention probabilities.
        max_position_embeddings (`int`, *optional*, defaults to 1026):
            The maximum sequence length that this model might ever be used with. Typically set this to something large
            just in case (e.g., 512 or 1024 or 2048).
        initializer_range (`float`, *optional*, defaults to 0.02):
            The standard deviation of the truncated_normal_initializer for initializing all weight matrices.
        layer_norm_eps (`float`, *optional*, defaults to 1e-12):
            The epsilon used by the layer normalization layers.
        position_embedding_type (`str`, *optional*, defaults to `"absolute"`):
            Type of position embedding. Choose one of `"absolute"`, `"relative_key"`, `"relative_key_query", "rotary"`.
            For positional embeddings use `"absolute"`. For more information on `"relative_key"`, please refer to
            [Self-Attention with Relative Position Representations (Shaw et al.)](https://arxiv.org/abs/1803.02155).
            For more information on `"relative_key_query"`, please refer to *Method 4* in [Improve Transformer Models
            with Better Relative Position Embeddings (Huang et al.)](https://arxiv.org/abs/2009.13658).
        is_decoder (`bool`, *optional*, defaults to `False`):
            Whether the model is used as a decoder or not. If `False`, the model is used as an encoder.
        use_cache (`bool`, *optional*, defaults to `True`):
            Whether or not the model should return the last key/values attentions (not used by all models). Only
            relevant if `config.is_decoder=True`.
        emb_layer_norm_before (`bool`, *optional*):
            Whether to apply layer normalization after embeddings but before the main stem of the network.
        token_dropout (`bool`, defaults to `False`):
            When this is enabled, masked tokens are treated as if they had been dropped out by input dropout.
    """
    model_type = "msat"

    def __init__(
        self,
        vocab_size=None,
        mask_token_id=None,
        pad_token_id=None,
        cls_token_id=None,
        eos_token_id=None,
        hidden_size=768,
        num_hidden_layers=12,
        num_attention_heads=12,
        intermediate_size=3072,
        hidden_dropout_prob=0.1,
        attention_probs_dropout_prob=0.1,
        max_position_embeddings=1026,
        max_position_embeddings_per_msa=128,
        initializer_range=0.02,
        layer_norm_eps=1e-12,
        position_embedding_type="absolute",
        use_cache=True,
        emb_layer_norm_before=None,
        token_dropout=False,
        vocab_list=None,
        **kwargs,
    ):
        super().__init__(pad_token_id=pad_token_id, mask_token_id=mask_token_id, **kwargs)

        self.vocab_size = vocab_size
        self.cls_token_id = cls_token_id
        self.eos_token_id = eos_token_id
        self.hidden_size = hidden_size
        self.num_hidden_layers = num_hidden_layers
        self.num_attention_heads = num_attention_heads
        self.intermediate_size = intermediate_size
        self.hidden_dropout_prob = hidden_dropout_prob
        self.attention_probs_dropout_prob = attention_probs_dropout_prob
        self.max_position_embeddings = max_position_embeddings
        self.max_position_embeddings_per_msa = max_position_embeddings_per_msa
        self.initializer_range = initializer_range
        self.layer_norm_eps = layer_norm_eps
        self.position_embedding_type = position_embedding_type
        self.use_cache = use_cache
        self.emb_layer_norm_before = emb_layer_norm_before
        self.token_dropout = token_dropout

        
class MSATPreTrainedModel(PreTrainedModel):
    """
    An abstract class to handle weights initialization and a simple interface for downloading and loading pretrained
    models.
    """

    config_class = MSATConfig
    base_model_prefix = "msat"
    _no_split_modules = []

    # Copied from transformers.models.bert.modeling_bert.BertPreTrainedModel._init_weights
    def _init_weights(self, module):
        """Initialize the weights"""
        if isinstance(module, nn.Linear):
            # Slightly different from the TF version which uses truncated_normal for initialization
            # cf https://github.com/pytorch/pytorch/pull/5617
            module.weight.data.normal_(mean=0.0, std=self.config.initializer_range)
            if module.bias is not None:
                module.bias.data.zero_()
        elif isinstance(module, nn.Embedding):
            module.weight.data.normal_(mean=0.0, std=self.config.initializer_range)
            if module.padding_idx is not None:
                module.weight.data[module.padding_idx].zero_()
        elif isinstance(module, nn.LayerNorm):
            module.bias.data.zero_()
            module.weight.data.fill_(1.0)

class MSATalphabet:
    """
     Mimic alphabet of MSA Transformer
    """
    def __init__(self, config):
        self.vocab_size = config.vocab_size
        self.padding_idx = config.pad_token_id
        self.mask_idx = config.mask_token_id
        self.cls_idx = config.cls_token_id
        self.eos_idx = config.eos_token_id
        self.prepend_bos = False
        self.append_eos = False
        
    
    def __len__(self):
        return self.vocab_size

class MSATargs:
    """
     Mimic args of MSA Transformer
    """
    def __init__(self, config):
        self.num_layers = config.num_hidden_layers
        self.layers = config.num_hidden_layers
        self.embed_dim = config.hidden_size
        self.login_bias = False
        self.ffn_embed_dim = config.intermediate_size
        self.attention_heads = config.num_attention_heads
        self.dropout = config.hidden_dropout_prob
        self.attention_dropout = config.attention_probs_dropout_prob
        self.activation_dropout = config.hidden_dropout_prob
        self.max_tokens_per_msa = config.max_position_embeddings_per_msa
        self.max_tokens = config.max_position_embeddings
        self.max_positions = config.max_position_embeddings_per_msa

class MSATransformerColContact(MSATransformer):
    def __init__(self, args, alphabet):
        super().__init__(args=args, alphabet=alphabet)
       
    def forward(self, tokens, repr_layers=[], need_head_weights=False, return_contacts=False):
        if return_contacts:
            need_head_weights = True
            
        result = super().forward(tokens=tokens, repr_layers= repr_layers, need_head_weights= need_head_weights, return_contacts=False)
        
        if return_contacts:
            # col_attentions: B x L x H x C x R x R
            # consider col_attentions of [CLS]
            attentions = result["col_attentions"][..., 0, :,:]
            batch_size, layers, heads, _ , seqlen, seqlen = result["col_attentions"].size()
            attentions = attentions.view(batch_size, layers, heads, seqlen, seqlen)

            contacts = self.contact_head(tokens, attentions)
            result["contacts"] = contacts
            
        return result
          
class MSATForContactPred(MSATPreTrainedModel):
    
    def __init__(self, config):
        super().__init__(config)
        self.num_labels = config.vocab_size
        self.config = config
        self.args = MSATargs(config)
        self.alphabet = MSATalphabet(config)
        self.msat = MSATransformerColContact(args=self.args,alphabet=self.alphabet)

        self.init_weights()
    
    def forward(
        self,
        input_ids: Optional[torch.LongTensor] = None,
        attention_mask: Optional[torch.Tensor] = None,
        position_ids: Optional[torch.LongTensor] = None,
        head_mask: Optional[torch.Tensor] = None,
        inputs_embeds: Optional[torch.FloatTensor] = None,
        labels: Optional[torch.LongTensor] = None,
        output_attentions: Optional[bool] = None,
        output_hidden_states: Optional[bool] = None,
        return_dict: Optional[bool] = None,
    ) -> Union[tuple, dict]:
        r"""
        labels (`torch.LongTensor` of shape `(batch_size, sequence_length)`, *optional*):
            Labels for computing the token classification loss. Indices should be in `[0, ..., config.num_labels - 1]`.
        """
        return_dict = return_dict if return_dict is not None else self.config.use_return_dict

        outputs = self.msat(
            tokens= input_ids,
            repr_layers= list(range(self.config.num_hidden_layers+1)),
            return_contacts = True,
        )

        logits = outputs["contacts"]

        loss = None
        if labels is not None:
            loss_fct = nn.BCELoss()

            labels = labels.to(device= logits.device, dtype= torch.float)
            loss = loss_fct(logits, labels)


        if not return_dict:
            output = (logits,) + outputs[2:]
            return ((loss,) + output) if loss is not None else output

        return TokenClassifierOutput(
            loss= loss,
            logits= logits,
            hidden_states= None,
            attentions= None,
        )