requirements.txt
python==3.10.13
numpy==1.23.5
torch==1.11.0
transformers==4.24.0
datasets==2.14.5
lingpy==2.6.10
scipy==1.11.3
scikit-learn==1.3.1
fair-esm==2.0.0
alphafold2_pytorch==0.4.0